// libs/admin-plugins/shipping/domestic-shipping/src/ui/generated-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
  Date: any;
};

export type Query = {
  __typename?: 'Query';
  administrators: AdministratorList;
  administrator?: Maybe<Administrator>;
  activeAdministrator?: Maybe<Administrator>;
  /** Get a list of Assets */
  assets: AssetList;
  /** Get a single Asset by id */
  asset?: Maybe<Asset>;
  me?: Maybe<CurrentUser>;
  channels: Array<Channel>;
  channel?: Maybe<Channel>;
  activeChannel: Channel;
  collections: CollectionList;
  /** Get a Collection either by id or slug. If neither id nor slug is speicified, an error will result. */
  collection?: Maybe<Collection>;
  collectionFilters: Array<ConfigurableOperationDefinition>;
  countries: CountryList;
  country?: Maybe<Country>;
  customerGroups: CustomerGroupList;
  customerGroup?: Maybe<CustomerGroup>;
  customers: CustomerList;
  customer?: Maybe<Customer>;
  facets: FacetList;
  facet?: Maybe<Facet>;
  globalSettings: GlobalSettings;
  job?: Maybe<Job>;
  jobs: JobList;
  jobsById: Array<Job>;
  jobQueues: Array<JobQueue>;
  order?: Maybe<Order>;
  orders: OrderList;
  paymentMethods: PaymentMethodList;
  paymentMethod?: Maybe<PaymentMethod>;
  productOptionGroups: Array<ProductOptionGroup>;
  productOptionGroup?: Maybe<ProductOptionGroup>;
  search: SearchResponse;
  products: ProductList;
  /** Get a Product either by id or slug. If neither id nor slug is speicified, an error will result. */
  product?: Maybe<Product>;
  /** Get a ProductVariant by id */
  productVariant?: Maybe<ProductVariant>;
  promotion?: Maybe<Promotion>;
  promotions: PromotionList;
  promotionConditions: Array<ConfigurableOperationDefinition>;
  promotionActions: Array<ConfigurableOperationDefinition>;
  roles: RoleList;
  role?: Maybe<Role>;
  shippingMethods: ShippingMethodList;
  shippingMethod?: Maybe<ShippingMethod>;
  shippingEligibilityCheckers: Array<ConfigurableOperationDefinition>;
  shippingCalculators: Array<ConfigurableOperationDefinition>;
  testShippingMethod: TestShippingMethodResult;
  testEligibleShippingMethods: Array<ShippingMethodQuote>;
  taxCategories: Array<TaxCategory>;
  taxCategory?: Maybe<TaxCategory>;
  taxRates: TaxRateList;
  taxRate?: Maybe<TaxRate>;
  zones: Array<Zone>;
  zone?: Maybe<Zone>;
  productReviews: ProductReviewList;
  productReview?: Maybe<ProductReview>;
  contacts: ContactList;
  contact?: Maybe<Contact>;
  subscribers: SubscriberList;
  subscriber?: Maybe<Subscriber>;
  newsletters: NewsletterList;
  newsletter?: Maybe<Newsletter>;
  productRecommendations: Array<ProductRecommendation>;
  productTierPrices: Array<TierPrice>;
  pincodes: PincodeList;
  pincode?: Maybe<Pincode>;
  checkPincode: Pincode;
  shippingCountry?: Maybe<ShippingCountryPrice>;
  shippingCountries: ShippingCountryPriceList;
  productGrid: ProductList;
  getOrderListByChannel: OrderList;
  getVendorByBrand?: Maybe<Vendor>;
  getVendorByEmail?: Maybe<Vendor>;
  vendors: VendorList;
  activeVendor?: Maybe<Vendor>;
  getVendorByID?: Maybe<Vendor>;
  dailyOrders: DailyOrderList;
  testDailyOrders?: Maybe<DailyOrderOutput>;
  delhiveryAccount?: Maybe<DelhiveryAccount>;
  delhiveryWarehouses: DelhiveryWarehouseList;
  delhiveryWarehouse?: Maybe<DelhiveryWarehouse>;
  delhiveryWarehouseByChannelId?: Maybe<DelhiveryWarehouse>;
  delhiveryWarehouseByPickupName?: Maybe<DelhiveryWarehouse>;
  getPackingSlip?: Maybe<PackingSlip>;
  ecomexpressAccountConfig?: Maybe<EcomExpressAccount>;
};
// ...
export type Collection = Node & {
  __typename?: 'Collection';
  isPrivate: Scalars['Boolean'];
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  languageCode?: Maybe<LanguageCode>;
  name: Scalars['String'];
  slug: Scalars['String'];
  breadcrumbs: Array<CollectionBreadcrumb>;
  position: Scalars['Int'];
  description: Scalars['String'];
  featuredAsset?: Maybe<Asset>;
  assets: Array<Asset>;
  parent?: Maybe<Collection>;
  children?: Maybe<Array<Collection>>;
  filters: Array<ConfigurableOperation>;
  translations: Array<CollectionTranslation>;
  productVariants: ProductVariantList;
  customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type CustomerGroupList = PaginatedList & {
  __typename?: 'CustomerGroupList';
  items: Array<CustomerGroup>;
  totalItems: Scalars['Int'];
};

export type CreateCustomerGroupInput = {
  name: Scalars['String'];
  customerIds?: Maybe<Array<Scalars['ID']>>;
};

export type UpdateCustomerGroupInput = {
  id: Scalars['ID'];
  name?: Maybe<Scalars['String']>;
};

export type Customer = Node & {
  __typename?: 'Customer';
  groups: Array<CustomerGroup>;
  history: HistoryEntryList;
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  title?: Maybe<Scalars['String']>;
  firstName: Scalars['String'];
  lastName: Scalars['String'];
  phoneNumber?: Maybe<Scalars['String']>;
  emailAddress: Scalars['String'];
  addresses?: Maybe<Array<Address>>;
  orders: OrderList;
  user?: Maybe<User>;
  customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type Facet = Node & {
  __typename?: 'Facet';
  isPrivate: Scalars['Boolean'];
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  languageCode: LanguageCode;
  name: Scalars['String'];
  code: Scalars['String'];
  values: Array<FacetValue>;
  translations: Array<FacetTranslation>;
  customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type JobList = PaginatedList & {
  __typename?: 'JobList';
  items: Array<Job>;
  totalItems: Scalars['Int'];
};
// ...
export type JobQueue = {
  __typename?: 'JobQueue';
  name: Scalars['String'];
  running: Scalars['Boolean'];
};
// ...
export type ConfigurableOperationDefinition = {
  __typename?: 'ConfigurableOperationDefinition';
  code: Scalars['String'];
  args: Array<ConfigArgDefinition>;
  description: Scalars['String'];
};
// ...
export type Node = {
  id: Scalars['ID'];
};

export enum SortOrder {
  Asc = 'ASC',
  Desc = 'DESC'
}
// ...
export type Administrator = Node & {
  __typename?: 'Administrator';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  firstName: Scalars['String'];
  lastName: Scalars['String'];
  emailAddress: Scalars['String'];
  user: User;
};

export type AdministratorList = PaginatedList & {
  __typename?: 'AdministratorList';
  items: Array<Administrator>;
  totalItems: Scalars['Int'];
};

export type Asset = Node & {
  __typename?: 'Asset';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  name: Scalars['String'];
  type: AssetType;
  fileSize: Scalars['Int'];
  mimeType: Scalars['String'];
  width: Scalars['Int'];
  height: Scalars['Int'];
  source: Scalars['String'];
  preview: Scalars['String'];
  focalPoint?: Maybe<Coordinate>;
};

export type Coordinate = {
  __typename?: 'Coordinate';
  x: Scalars['Float'];
  y: Scalars['Float'];
};

export type AssetList = PaginatedList & {
  __typename?: 'AssetList';
  items: Array<Asset>;
  totalItems: Scalars['Int'];
};

export enum AssetType {
  Image = 'IMAGE',
  Video = 'VIDEO',
  Binary = 'BINARY'
}

export type CurrentUser = {
  __typename?: 'CurrentUser';
  id: Scalars['ID'];
  identifier: Scalars['String'];
  channels: Array<CurrentUserChannel>;
};

export type CurrentUserChannel = {
  __typename?: 'CurrentUserChannel';
  id: Scalars['ID'];
  token: Scalars['String'];
  code: Scalars['String'];
  permissions: Array<Permission>;
};

export type Channel = Node & {
  __typename?: 'Channel';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  code: Scalars['String'];
  token: Scalars['String'];
  defaultTaxZone?: Maybe<Zone>;
  defaultShippingZone?: Maybe<Zone>;
  defaultLanguageCode: LanguageCode;
  currencyCode: CurrencyCode;
  pricesIncludeTax: Scalars['Boolean'];
  shippingMethod: Array<ShippingMethod>;
};
// ...
export type CollectionList = PaginatedList & {
  __typename?: 'CollectionList';
  items: Array<Collection>;
  totalItems: Scalars['Int'];
};

export type ProductVariantList = PaginatedList & {
  __typename?: 'ProductVariantList';
  items: Array<ProductVariant>;
  totalItems: Scalars['Int'];
};

export type Country = Node & {
  __typename?: 'Country';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  languageCode: LanguageCode;
  code: Scalars['String'];
  name: Scalars['String'];
  enabled: Scalars['Boolean'];
  translations: Array<CountryTranslation>;
};

export type CountryTranslation = {
  __typename?: 'CountryTranslation';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  languageCode: LanguageCode;
  name: Scalars['String'];
};

export type CountryList = PaginatedList & {
  __typename?: 'CountryList';
  items: Array<Country>;
  totalItems: Scalars['Int'];
};

export type CustomerGroup = Node & {
  __typename?: 'CustomerGroup';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  name: Scalars['String'];
  customers: CustomerList;
};


export type CustomerGroupCustomersArgs = {
  options?: Maybe<CustomerListOptions>;
};

export type CustomerList = PaginatedList & {
  __typename?: 'CustomerList';
  items: Array<Customer>;
  totalItems: Scalars['Int'];
};
// ...
export type FacetList = PaginatedList & {
  __typename?: 'FacetList';
  items: Array<Facet>;
  totalItems: Scalars['Int'];
};

export type GlobalSettings = {
  __typename?: 'GlobalSettings';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  availableLanguages: Array<LanguageCode>;
  trackInventory: Scalars['Boolean'];
  outOfStockThreshold: Scalars['Int'];
  serverConfig: ServerConfig;
  customFields?: Maybe<GlobalSettingsCustomFields>;
};
// ...
export type Pincode = Node & {
  __typename?: 'Pincode';
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  id: Scalars['ID'];
  pincode?: Maybe<Scalars['Int']>;
  state?: Maybe<Scalars['String']>;
  district?: Maybe<Scalars['String']>;
  prepaid?: Maybe<Scalars['Boolean']>;
  cod?: Maybe<Scalars['Boolean']>;
  pickup?: Maybe<Scalars['Boolean']>;
  cash?: Maybe<Scalars['Boolean']>;
  repl?: Maybe<Scalars['Boolean']>;
};

export type PincodeList = PaginatedList & {
  __typename?: 'PincodeList';
  items: Array<Pincode>;
  totalItems: Scalars['Int'];
};

export type PincodeDeleteReturn = {
  __typename?: 'PincodeDeleteReturn';
  error?: Maybe<Scalars['String']>;
  message?: Maybe<Scalars['String']>;
};

export type PincodeListOptions = {
  skip: Scalars['Int'];
  take: Scalars['Int'];
  sort?: Maybe<PincodeSortParameter>;
  filter?: Maybe<PincodeFilterParameter>;
};
// ...
export type DailyOrderOutput = {
  __typename?: 'DailyOrderOutput';
  items: Array<OrdersPerMonth>;
};
// ...
export namespace Pincode {
  export type Fragment = PincodeFragment;
}

export namespace PincodeDeleteReturn {
  export type Fragment = PincodeDeleteReturnFragment;
}

export namespace GetAllPincode {
  export type Variables = GetAllPincodeQueryVariables;
  export type Query = GetAllPincodeQuery;
  export type Pincodes = (NonNullable<GetAllPincodeQuery['pincodes']>);
  export type Items = NonNullable<(NonNullable<(NonNullable<GetAllPincodeQuery['pincodes']>)['items']>)[number]>;
}

export namespace DeletePincode {
  export type Variables = DeletePincodeMutationVariables;
  export type Mutation = DeletePincodeMutation;
  export type DeletePincode = (NonNullable<DeletePincodeMutation['deletePincode']>);
}

export namespace CheckPincode {
  export type Variables = CheckPincodeQueryVariables;
  export type Query = CheckPincodeQuery;
  export type CheckPincode = (NonNullable<CheckPincodeQuery['checkPincode']>);
}
// ...
export type PincodeFragment = (
  { __typename?: 'Pincode' }
  & Pick<Pincode, 'id' | 'createdAt' | 'updatedAt' | 'pincode' | 'state' | 'district' | 'prepaid' | 'cod' | 'pickup' | 'cash' | 'repl'>
);

export type PincodeDeleteReturnFragment = (
  { __typename?: 'PincodeDeleteReturn' }
  & Pick<PincodeDeleteReturn, 'error' | 'message'>
);

export type GetAllPincodeQueryVariables = Exact<{
  options?: Maybe<PincodeListOptions>;
}>;


export type GetAllPincodeQuery = (
  { __typename?: 'Query' }
  & { pincodes: (
    { __typename?: 'PincodeList' }
    & Pick<PincodeList, 'totalItems'>
    & { items: Array<(
      { __typename?: 'Pincode' }
      & PincodeFragment
    )> }
  ) }
);

export type DeletePincodeMutationVariables = Exact<{
  id: Scalars['ID'];
}>;


export type DeletePincodeMutation = (
  { __typename?: 'Mutation' }
  & { deletePincode: (
    { __typename?: 'PincodeDeleteReturn' }
    & PincodeDeleteReturnFragment
  ) }
);

export type CheckPincodeQueryVariables = Exact<{
  pincode: CheckPincodeInput;
}>;


export type CheckPincodeQuery = (
  { __typename?: 'Query' }
  & { checkPincode: (
    { __typename?: 'Pincode' }
    & PincodeFragment
  ) }
);